analytics-service/requirements.txt
Django==3.2.12
psycopg2-binary==2.9.3
django-health-check==3.16.4
djangorestframework==3.12.4
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-environ==0.8.1
djangorestframework-simplejwt==5.0.0
djangorestframework>=3.12,<4.0
requests>=2.25.1,<3.0

